rename gh60 file to revc
