+++ keyboards/gh60/revc/revc.c
@@ -1,4 +1,4 @@
-#include "gh60.h"
+#include "revc.h"
 
 
 extern inline void gh60_caps_led_on(void);
